# backendRL/src/main/java/org/example/backendrl/config/api-docs.yaml
openapi: 3.1.0
info:
  title: RecordingLogs API
  description: "Music rating and recommendation platform: genres, artists, albums,\
    \ songs, ratings, reviews, album lists, favorites and follows."
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/songs/{id}:
    get:
      tags:
      - Songs
      summary: Get a song by id
      operationId: getSong
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: Song found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SongResponse"
    put:
      tags:
      - Songs
      summary: Update a song
      operationId: updateSong
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongRequest"
        required: true
      responses:
        "404":
          description: Song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: Song updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SongResponse"
    delete:
      tags:
      - Songs
      summary: Delete a song
      operationId: deleteSong
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Song deleted
...
  /api/songs:
    get:
      tags:
      - Songs
      summary: Get all songs
      operationId: getAllSongs
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of songs
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SongResponse"
    post:
      tags:
      - Songs
      summary: Create a song
      operationId: createSong
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Song created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SongResponse"
...
  /api/artists/{artistId}/songs/{songId}:
    post:
      tags:
      - Artists
      summary: Add a song to an artist
      operationId: addSongToArtist
      parameters:
      - name: artistId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Artist or song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Song is already linked to this artist
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Song added to artist
    delete:
      tags:
      - Artists
      summary: Remove a song from an artist
      operationId: removeSongFromArtist
      parameters:
      - name: artistId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Artist, song or link not found"
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Song removed from artist
...
  /api/albums/{albumId}/songs/{songId}:
    post:
      tags:
      - Albums
      summary: Add a song to an album's tracklist
      operationId: addSongToAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumSongRequest"
        required: true
      responses:
        "404":
          description: Album or song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Song is already on this album
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "201":
          description: Song added to album
    delete:
      tags:
      - Albums
      summary: Remove a song from an album's tracklist
      operationId: removeSongFromAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Album, song or link not found"
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "204":
          description: Song removed from album
...
  /api/songs/{songId}/artists:
    get:
      tags:
      - Songs
      summary: Get all artists of a song
      operationId: getArtistsForSong
      parameters:
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of artists
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistSummaryResponse"
  /api/songs/{songId}/albums:
    get:
      tags:
      - Songs
      summary: Get all albums containing a song
      operationId: getAlbumsForSong
      parameters:
      - name: songId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Song not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of albums
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSummaryResponse"
...
  /api/artists/{artistId}/songs:
    get:
      tags:
      - Artists
      summary: Get all songs of an artist
      operationId: getSongsForArtist
      parameters:
      - name: artistId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Artist not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of songs
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SongSummaryResponse"
...
  /api/albums/{albumId}/songs:
    get:
      tags:
      - Albums
      summary: Get the tracklist of an album
      operationId: getSongsForAlbum
      parameters:
      - name: albumId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Album not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "422":
          description: Unprocessable Content
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        "200":
          description: List of songs
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumSongResponse"
...
components:
  schemas:
...
    SongRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        durationSeconds:
          type: integer
          format: int32
        originalReleaseDate:
          type: string
          format: date
      required:
      - title
    SongResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        durationSeconds:
          type: integer
          format: int32
        originalReleaseDate:
          type: string
          format: date
...
    AlbumSongRequest:
      type: object
      properties:
        trackNumber:
          type: integer
          format: int32
          minimum: 1
        discNumber:
          type: integer
          format: int32
          minimum: 1
      required:
      - trackNumber
...
    ArtistSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    AlbumSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        releaseDate:
          type: string
          format: date
        coverUrl:
          type: string
    SongSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        durationSeconds:
          type: integer
          format: int32
        originalReleaseDate:
          type: string
          format: date
    AlbumSongResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        albumId:
          type: integer
          format: int64
        songId:
          type: integer
          format: int64
        songTitle:
          type: string
        trackNumber:
          type: integer
          format: int32
        discNumber:
          type: integer
          format: int32